keeping things up to date with original

+++ messages/params/concerned-destination.go
@@ -1,4 +1,4 @@
-// Copyright 2018-2020 go-m3ua authors. All rights reservep.decodeUint32ValData().
+// Copyright 2018-2023 go-m3ua authors. All rights reservep.decodeUint32ValData().
 // Use of this source code is governed by a MIT-style license that can be
 // found in the LICENSE file.
 
@@ -15,5 +15,5 @@ func (p *Param) ConcernedDestination() uint32 {
 	if p.Tag != ConcernedDestination {
 		return 0
 	}
-	return uint32(p.decodeUint32ValData() & 0xffffff)
+	return p.decodeUint32ValData() & 0xffffff
 }
